TestSetAccess abstraction as preparation for Android support
