+++ src/main/kotlin/com/code_intelligence/cifuzz/CIFuzzExtension.kt
@@ -19,4 +19,9 @@ interface CIFuzzExtension {
      * Configures the task that executes the fuzz tests - defaults to the task with the same name as the source set.
      */
     val testTask: Property<Test>
+
+    /**
+     * Configure the source set that contains the production code under tests – defaults to 'sourceSets.main'.
+     */
+    val mainSourceSet: Property<SourceSet>
 }
